pnfs: fix pnfs_{read|write}pages
THIS WILL BREAK OTHER LAYOUT DRIVERS that rely on old behavior.

This patch fixes two basic problems:
1. It is impossible to fall back to nfs from the layout driver's
read_pagelist or write_pagelist routines.
2. The pnfs.c code is assuming that if these routines return a non-zero
status, that pnfs_writeback_done (or pnfs_read_done) was not called, which
is not consistent with them returning count of bytes read/written.

This patch alters the call chain, and expects
ld_io_ops->{read|write}_pagelist to return either a 0 or 1, consistent
with its parents, putting any other information that needs to be returned
into the data.res and data.task structures.

+++ include/linux/nfs4_pnfs.h
@@ -116,8 +116,14 @@ struct layoutdriver_io_operations {
 	/* Functions that use the pagecache.
 	 * If use_pagecache == 1, then these functions must be implemented.
 	 */
-	ssize_t (*read_pagelist) (struct pnfs_layout_type *layoutid, struct page **pages, unsigned int pgbase, unsigned nr_pages, loff_t offset, size_t count, struct nfs_read_data *nfs_data);
-	ssize_t (*write_pagelist) (struct pnfs_layout_type *layoutid, struct page **pages, unsigned int pgbase, unsigned nr_pages, loff_t offset, size_t count, int sync, struct nfs_write_data *nfs_data);
+	int (*read_pagelist) (struct pnfs_layout_type *layoutid,
+			      struct page **pages, unsigned int pgbase,
+			      unsigned nr_pages, loff_t offset, size_t count,
+			      struct nfs_read_data *nfs_data);
+	int (*write_pagelist) (struct pnfs_layout_type *layoutid,
+			       struct page **pages, unsigned int pgbase,
+			       unsigned nr_pages, loff_t offset, size_t count,
+			       int sync, struct nfs_write_data *nfs_data);
 	int (*flush_one) (struct inode *inode, struct list_head *head, unsigned int npages, size_t count, int how);
 	void (*free_request_data) (struct nfs_page *);
